refactor: holeData読み込みをuseHoleDataカスタムフックに統一
