+++ components/greens/GreenCardPDF.tsx
@@ -14,6 +14,7 @@ import {
   scalePathToPixels,
   ydToPx,
 } from "@/lib/greenCanvas.convert";
+import { useHoleData } from "@/hooks/useHoleData";
 
 interface Props {
   hole: string;
@@ -26,15 +27,7 @@ const DEPTH_FONT_SIZE = 80;
 const HORIZONTAL_FONT_SIZE = 40;
 
 export default function GreenCardPDF({ hole, currentPin }: Props) {
-  const [holeData, setHoleData] = useState<HoleData | null>(null);
-
-  useEffect(() => {
-    const paddedHole = hole.padStart(2, "0");
-    fetch(`/greens/hole_${paddedHole}.json`)
-      .then((res) => res.json())
-      .then((data) => setHoleData(data))
-      .catch((err) => console.error("JSON読み込みエラー:", err));
-  }, [hole]);
+  const holeData = useHoleData(hole);
 
   if (!holeData) {
     return <div>読み込み中...</div>;
