+++ components/greens/GreenCardPDFExport.tsx
@@ -9,7 +9,12 @@ import {
   getBoundaryIntersectionX,
   getBoundaryIntersectionY,
 } from "@/lib/greenCanvas.geometry";
-import { CANVAS_SIZE, scalePathToPixels, ydToPx } from "@/lib/greenCanvas.convert";
+import {
+  CANVAS_SIZE,
+  scalePathToPixels,
+  ydToPx,
+} from "@/lib/greenCanvas.convert";
+import { useHoleData } from "@/hooks/useHoleData";
 
 interface Props {
   hole: string;
@@ -46,15 +51,7 @@ export default function GreenCardPDFExport({
   height = 240,
   currentPin,
 }: Props) {
-  const [holeData, setHoleData] = useState<HoleData | null>(null);
-
-  useEffect(() => {
-    const paddedHole = hole.padStart(2, "0");
-    fetch(`/greens/hole_${paddedHole}.json`)
-      .then((res) => res.json())
-      .then((data) => setHoleData(data))
-      .catch((err) => console.error("JSON読み込みエラー:", err));
-  }, [hole]);
+  const holeData = useHoleData(hole);
 
   if (!holeData) {
     return <div>読み込み中...</div>;
